# Define attributes for chat message model