--- a/app/Models/Chat/Message.php
+++ b/app/Models/Chat/Message.php
@@ -43,4 +43,23 @@ class Message extends Model
     {
         return $query->where('message_id', '>', $messageId);
     }
+
+    public function getAttribute($key)
+    {
+        return match ($key) {
+            'channel_id',
+            'content',
+            'message_id',
+            'user_id' => $this->getRawAttribute($key),
+
+            'is_action' => (bool) $this->getRawAttribute($key),
+
+            'timestamp' => $this->getTimeFast($key),
+
+            'timestamp_json' => $this->getJsonTimeFast($key),
+
+            'channel',
+            'sender' => $this->getRelationValue($key),
+        };
+    }
 }