--- a/app/Transformers/Chat/MessageTransformer.php
+++ b/app/Transformers/Chat/MessageTransformer.php
@@ -22,7 +22,7 @@ class MessageTransformer extends TransformerAbstract
             'message_id' => $message->message_id,
             'sender_id' => $message->user_id,
             'channel_id' => $message->channel_id,
-            'timestamp' => json_time($message->timestamp),
+            'timestamp' => $message->timestamp_json,
             'content' => $message->content,
             'is_action' => $message->is_action,
         ];